+++ TextString.h
@@ -1,17 +1,18 @@
 #ifndef TEXTSTRING_H
 #define TEXTSTRING_H
 
+#include <stdio.h>
 #include <stdbool.h>
 typedef struct {
   char *string;
   bool hasEndOfString;
-  int size; //including end of string if it has
+  size_t size; //including end of string if it has
 } TextStringType;
 
 
-void TxtStr_storeTextString(TextStringType *, char *, int);
+void TxtStr_storeTextString(TextStringType *, char *, size_t);
 
-int TxtStr_getStringLen(TextStringType);
+size_t TxtStr_getStringLen(TextStringType);
 
 void TxtStr_freeTextString(TextStringType *);
